+++ spec/tests/ContentMetadata.spec.ts
@@ -1,5 +1,5 @@
 /// <reference path='../../src/ts/Conviva.d.ts'/>
-import {PlayerType, StreamType, VRContentType} from 'bitmovin-player';
+import { PlayerType, StreamType, VRContentType } from 'bitmovin-player';
 import { ConvivaAnalytics } from '../../src/ts';
 import { MockHelper, TestingPlayerAPI } from '../helper/MockHelper';
 
@@ -34,7 +34,7 @@ describe('content metadata spec', () => {
       }));
     });
 
-    it('default playerType in custom tags', () => {
+    it('sets the default playerType in custom tags', () => {
       jest.spyOn(playerMock, 'getPlayerType').mockReturnValue(PlayerType.Native);
       playerMock.eventEmitter.firePlayEvent();
 
@@ -43,7 +43,7 @@ describe('content metadata spec', () => {
       }));
     });
 
-    it('default streamType in custom tags', () => {
+    it('sets the default streamType in custom tags', () => {
       jest.spyOn(playerMock, 'getStreamType').mockReturnValue(StreamType.Dash);
       playerMock.eventEmitter.firePlayEvent();
 
@@ -52,7 +52,7 @@ describe('content metadata spec', () => {
       }));
     });
 
-    it('default vrContentType in custom tags', () => {
+    it('sets the default vrContentType in custom tags', () => {
       playerMock.eventEmitter.firePlayEvent();
 
       expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(expect.objectContaining({
@@ -63,7 +63,7 @@ describe('content metadata spec', () => {
     it('override playerType in custom tags', () => {
       jest.spyOn(playerMock, 'getPlayerType').mockReturnValue(PlayerType.Native);
       jest.spyOn(playerMock, 'getStreamType').mockReturnValue(StreamType.Dash);
-      convivaAnalytics.updateContentMetadata({ custom: { myTag: 'withMyValue', playerType: PlayerType.Html5 }, assetName: 'MyAsset' });
+      convivaAnalytics.updateContentMetadata({ custom: { playerType: PlayerType.Html5 }, assetName: 'MyAsset' });
       playerMock.eventEmitter.firePlayEvent();
 
       expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(expect.objectContaining({
@@ -74,7 +74,7 @@ describe('content metadata spec', () => {
     it('override streamType in custom tags', () => {
       jest.spyOn(playerMock, 'getPlayerType').mockReturnValue(PlayerType.Native);
       jest.spyOn(playerMock, 'getStreamType').mockReturnValue(StreamType.Dash);
-      convivaAnalytics.updateContentMetadata({ custom: { myTag: 'withMyValue', streamType: 'dash_vod' }, assetName: 'MyAsset' });
+      convivaAnalytics.updateContentMetadata({ custom: { streamType: 'dash_vod' }, assetName: 'MyAsset' });
       playerMock.eventEmitter.firePlayEvent();
 
       expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(expect.objectContaining({
@@ -83,7 +83,7 @@ describe('content metadata spec', () => {
     });
 
     it('override vrContentType in custom tags', () => {
-      convivaAnalytics.updateContentMetadata({ custom: { myTag: 'withMyValue', vrContentType: VRContentType.Single }, assetName: 'MyAsset' });
+      convivaAnalytics.updateContentMetadata({ custom: { vrContentType: VRContentType.Single }, assetName: 'MyAsset' });
       playerMock.eventEmitter.firePlayEvent();
 
       expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(expect.objectContaining({